Correccion de erorres y se agregaron las funciones faltantes

+++ Laboratorio 6/Atom/Lectora.h
@@ -20,7 +20,7 @@ bool vacioArch = false; //.m
 class Lectora {
     private:
       fstream archivo;
-      int xK;
+      float xK;
       int renglon;
       int totales;
